Annual Sales of Cars in Europe
Source: R/data_eurocars.R

Annual model-level sales of cars in European countries between 1970 and 1999.

data('eurocars')

Format

A data.frame with 11,549 observations on 39 variables:

year: Year of the observation

market_id: Numeric identifier for geographical market

model_id: Numeric identifier for the car

model_id2: Numeric identifier for the car with unified products when model number changes

type: Car Type

brand: Brand of Manufacturer

model: Car Model

origin_id: Numeric identifier for country of origin

location_id: Numeric identifier for location

class_id: Numeric vehicle class identifier

domestic: == 1 if produced domestically

firm_id: Numeric identifier for firm that produced vehicle

qty_sold: Quantity of cars sold

displacement: Engine capacity in cc

kilowatts: Engine Power in kilowatts

weight: Vehicle weight in kilograms

n_seats: Number of seats

n_doors: Number of doors

length: Vehicle length in cm

height: Vehicle height in cm

fueleff_90: Fuel Efficiency at 90 km/hr

fueleff_90: Fuel Efficiency at 120 km/hr

fueleff_city: Fuel Efficiency when driving in city

fueleff_avg: Average Fuel Efficiency

max_speed: Maximum speed in km/hr

accel: Time taken in seconds from 0 to 100 km/hr

price: Price in local currency

price_euro: Price in Euros

exchrate_exp: Exchange rate in exporting country

exchrate_imp: Exchange rate in importing country

cpi_exp: Consumer Price Index in exporting country

ppi_exp: Producer Price Index in exporting country

cpi_imp: Consumer Price Index in importing country

ppi_imp: Producer Price Index in importing country

tax: Tax rate on purchase

population: Country of sale's population

ngdp: Nominal GDP in country of sale

rgdp: Real GDP in country of sale

Details

Data Source: Brenkers and Verboven, 2006, "Liberalizing a Distribution System: The European Car Market", Journal of the European Economic Association

Examples

 str(eurocars)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	11549 obs. of  39 variables:
#>  $ year        : num  83 84 85 86 87 88 89 90 91 92 ...
#>  $ market_id   : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ model_id    : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ model_id2   : num  5 5 5 5 5 5 5 5 5 5 ...
#>  $ type        : chr  "alfa 33" "alfa 33" "alfa 33" "alfa 33" ...
#>  $ brand       : chr  "alfa romeo" "alfa romeo" "alfa romeo" "alfa romeo" ...
#>  $ model       : chr  "33" "33" "33" "33" ...
#>  $ origin_id   : num  3 3 3 3 3 3 3 3 3 3 ...
#>  $ location_id : num  5 5 5 5 5 5 5 5 5 5 ...
#>  $ class_id    : num  2 2 2 2 2 2 2 2 2 2 ...
#>  $ domestic    : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ firm_id     : num  1 1 1 1 4 4 4 4 4 4 ...
#>  $ qty_sold    : num  729 1860 1771 2047 2147 ...
#>  $ displacement: num  1351 1351 1351 1351 1351 ...
#>  $ kilowatts   : num  58 58 58 58 58 58 63 63 65 66 ...
#>  $ weight      : num  890 890 890 890 910 910 910 940 940 940 ...
#>  $ n_seats     : num  5 5 5 5 5 5 5 5 5 5 ...
#>  $ n_doors     : num  5 5 5 5 5 5 5 5 5 5 ...
#>  $ length      : num  402 402 402 402 402 ...
#>  $ width       : num  161 161 161 161 161 ...
#>  $ height      : num  130 130 130 134 134 ...
#>  $ fueleff_90  : num  5.8 5.8 5.8 5.8 5.8 ...
#>  $ fueleff_120 : num  7.9 7.9 7.9 7.9 8 ...
#>  $ fueleff_city: num  9.6 9.6 9.6 9.6 10 ...
#>  $ fueleff_avg : num  7.77 7.77 7.77 7.77 7.93 ...
#>  $ max_speed   : num  165 165 165 167 167 167 172 176 176 178 ...
#>  $ accel       : num  NA 11.5 11.5 11.5 11.5 ...
#>  $ price       : num  336250 348750 361000 339900 349900 ...
#>  $ price_euro  : num  6078 5859 6126 6559 7272 ...
#>  $ exchrate_exp: num  1649 1818 1916 1733 1676 ...
#>  $ exchrate_imp: num  55.3 59.5 58.9 51.8 48.1 ...
#>  $ cpi_exp     : num  62.7 69.5 75.9 80.3 84.1 ...
#>  $ ppi_exp     : num  71.5 78.8 84.9 85 87.6 ...
#>  $ cpi_imp     : num  80.8 85.9 90.1 91.3 92.7 ...
#>  $ ppi_imp     : num  100.4 108 110.8 98.1 93.3 ...
#>  $ tax         : num  0.25 0.25 0.25 0.25 0.25 ...
#>  $ population  : num  9860000 9860000 9860000 9860000 9870000 ...
#>  $ ngdp        : num  4.19e+12 4.51e+12 4.83e+12 5.08e+12 5.32e+12 ...
#>  $ rgdp        : num  5.45e+12 5.59e+12 5.64e+12 5.73e+12 5.86e+12 ...